requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
exceptiongroup==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
kombu==5.2.4
msgpack==1.0.4
packaging==23.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
